LEGACY: Interlude server update: 	* Updated libraries: 		1. C3P0 to 0.9.2-pre1. 		1. Javolution to 5.5.1 and related code changes. 		1. MMOCore minor update to use new Javolution. 		1. Added MChange-Commons v0.2, required by new C3P0. 		1. MySQL-Connector to 5.1.18. 	* Refactored random, using current High Five version, removed MTRandom. 	* Using same warning/error than High Five. 	* Fixed all warnings and errors. 	* Fixed NPEs. 	* Fixed some bad code. 	* Ant Buildier needs some fixes, anyway was updated a bit as well. '''Note 1:''' Interlude is '''not''' being actively developed, this update is to make it compatible with current software (Java 7, MySQL, etc), in other words ''usable''.
'''Note 2:''' For Interlude __'''always'''__ use Legacy forums!

+++ java/net/sf/l2j/gameserver/serverpackets/RecipeShopItemInfo.java
@@ -50,7 +50,7 @@ public class RecipeShopItemInfo  extends L2GameServerPacket
 		writeD(_shopId);
 		writeD(_recipeId);
 		writeD(manufacturer != null ? (int)manufacturer.getCurrentMp() : 0);
-		writeD(manufacturer != null ? (int)manufacturer.getMaxMp() : 0);
+		writeD(manufacturer != null ? manufacturer.getMaxMp() : 0);
 		writeD(0xffffffff);
 	}
